requirements.txt
numpy
Pillow
torchvision
onnxruntime-gpu
scipy
tqdm